+++ src/Generators/Factories/PropertyFactory.php
@@ -12,6 +12,7 @@ use PhpUnitGen\Core\Contracts\Aware\ImportFactoryAware;
 use PhpUnitGen\Core\Contracts\Aware\MockGeneratorAware;
 use PhpUnitGen\Core\Contracts\Generators\Factories\PropertyFactory as PropertyFactoryContract;
 use PhpUnitGen\Core\Generators\Concerns\InstantiatesClass;
+use PhpUnitGen\Core\Helpers\Reflect;
 use PhpUnitGen\Core\Models\TestClass;
 use PhpUnitGen\Core\Models\TestImport;
 use PhpUnitGen\Core\Models\TestProperty;
@@ -62,9 +63,10 @@ class PropertyFactory implements
         $stringType = 'mixed';
         $isBuiltIn = true;
 
-        if ($reflectionParameter->getType()) {
-            $stringType = strval($reflectionParameter->getType());
-            $isBuiltIn = $reflectionParameter->getType()->isBuiltin();
+        $reflectionType = Reflect::parameterType($reflectionParameter);
+        if ($reflectionType) {
+            $stringType = $reflectionType->getType();
+            $isBuiltIn = $reflectionType->isBuiltin();
         }
 
         return $this->makeCustom($class, $reflectionParameter->getName(), $stringType, $isBuiltIn);